+++ geofileops/geoops.py
@@ -2,18 +2,22 @@
 Module exposing all supported operations on geomatries in geofiles.
 """
 
+from datetime import datetime
 import logging
 import logging.config
 from pathlib import Path
+import shutil
 from typing import Any, Callable, List, Literal, Optional, Tuple, Union, TYPE_CHECKING
 import warnings
 
 from pygeoops import GeometryType
 import shapely
 
+from geofileops import fileops
 from geofileops.util import _geoops_gpd
 from geofileops.util import _geoops_sql
 from geofileops.util import _geoops_ogr
+from geofileops.util import _io_util
 from geofileops.util._geometry_util import (
     BufferEndCapStyle,
     BufferJoinStyle,
@@ -24,6 +28,310 @@ if TYPE_CHECKING:
     import os
 
 logger = logging.getLogger(__name__)
+
+
+def dissolve_within_distance(
+    input_path: Path,
+    output_path: Path,
+    distance: float,
+    gridsize: float,
+    close_internal_gaps: bool = False,
+    input_layer: Optional[str] = None,
+    output_layer: Optional[str] = None,
+    nb_parallel: int = -1,
+    batchsize: int = -1,
+    force: bool = False,
+):
+    """
+    Dissolve geometries that are within the distance specified.
+
+    The output layer will contain the dissolved geometries where all gaps between the
+    input geometries up to `distance` are closed.
+
+    Notes:
+      - Only tested on polygon input.
+      - Gaps between the individual polygons of multipolygon input features will also
+        be closed.
+      - The polygons in the output file are exploded to simple geometries.
+      - No attributes from the input layer are retained.
+      - If `close_internal_gaps` is False, the default, a `gridsize` > 0
+        (E.g. 0.000001) should be specified, otherwise some input boundary gaps could
+        still be closed due to rounding side effects.
+
+    Alternative names:
+      - ArcMap: aggregate_polygons (similar functionality)
+      - Keywords: merge, dissolve, aggregate, snap, close gaps, union
+
+    Args:
+        input_path (PathLike): the input file.
+        output_path (PathLike): the file to write the result to.
+        distance (float): the maximum distance between geometries to be dissolved.
+        gridsize (float, optional): the size of the grid the coordinates of the ouput
+            will be rounded to. Eg. 0.001 to keep 3 decimals. Value 0.0 doesn't change
+            the precision. If `close_boundary_gaps` is False, the default, a
+            `gridsize` > 0 (E.g. 0.000001) should be specified, otherwise some boundary
+            gaps in the input geometries could still be closed due to rounding side
+            effects.
+        close_internal_gaps (bool, optional): also close gaps, strips or holes in the
+            input geometries that are narrower than the `distance` specified. E.g. small
+            holes, narrow strips starting at the boundary,... Defaults to False.
+        input_layer (str, optional): input layer name. Optional if the input
+            file only contains one layer.
+        output_layer (str, optional): input layer name. Optional if the input
+            file only contains one layer.
+        nb_parallel (int, optional): the number of parallel processes to use.
+            Defaults to -1: use all available CPUs.
+        batchsize (int, optional): indicative number of rows to process per
+            batch. A smaller batch size, possibly in combination with a
+            smaller nb_parallel, will reduce the memory usage.
+            Defaults to -1: (try to) determine optimal size automatically.
+        force (bool, optional): overwrite existing output file(s).
+            Defaults to False.
+    """
+    start_time = datetime.now()
+    operation_name = "dissolve_within_distance"
+    logger = logging.getLogger(f"geofileops.{operation_name}")
+    nb_steps = 4
+    if not close_internal_gaps:
+        # 3 extra steps if boundary gaps not to be closed.
+        nb_steps += 3
+
+    # Already check here if it is useful to continue
+    if output_path.exists():
+        if force is False:
+            logger.info(f"Stop, output exists already {output_path}")
+            return
+        else:
+            fileops.remove(output_path)
+
+    tempdir = _io_util.create_tempdir(f"geofileops/{operation_name}")
+    try:
+        # First dissolve the input.
+        #
+        # Note: this reduces the complexity of operations to be executed later on.
+        # Note2: this already applies the gridsize, which needs to be applied anyway to
+        # avoid issues when determining the addedpieces_1neighbour later on.
+        logger.info(f"Start, with input file {input_path}")
+        step = 1
+        logger.info(f"Step {step} of {nb_steps}")
+        diss_path = tempdir / "100_diss.gpkg"
+        _geoops_gpd.dissolve(
+            input_path=input_path,
+            output_path=diss_path,
+            explodecollections=True,
+            input_layer=input_layer,
+            gridsize=gridsize,
+            nb_parallel=nb_parallel,
+            batchsize=batchsize,
+            operation_prefix=f"{operation_name}-",
+        )
+
+        # Positive buffer of distance / 2 to close all gaps.
+        #
+        # Note: gridsize is not applied to preserve all possible accuracy for these
+        # temporary boundaries, otherwise the polygons are sometimes enlarged slightly,
+        # which isn't wanted + creates issues when determining the
+        # addedpieces_1neighbour later on.
+        step += 1
+        logger.info(f"Step {step} of {nb_steps}")
+        buff_path = tempdir / "110_diss_bufp.gpkg"
+        _geoops_gpd.buffer(
+            input_path=diss_path,
+            output_path=buff_path,
+            distance=distance / 2,
+            endcap_style=BufferEndCapStyle.SQUARE,
+            join_style=BufferJoinStyle.MITRE,
+            mitre_limit=1.25,
+            # gridsize=gridsize,
+            nb_parallel=nb_parallel,
+            batchsize=batchsize,
+            operation_prefix=f"{operation_name}-",
+        )
+
+        # Dissolve the buffered input
+        #
+        # Note: gridsize is not applied to preserve all possible accuracy for these
+        # temporary boundaries, otherwise the polygons are sometimes enlarged slightly,
+        # which isn't wanted + creates issues when determining the
+        # addedpieces_1neighbour later on.
+        step += 1
+        logger.info(f"Step {step} of {nb_steps}")
+        buff_diss_path = tempdir / "120_diss_bufp_diss.gpkg"
+        _geoops_gpd.dissolve(
+            input_path=buff_path,
+            output_path=buff_diss_path,
+            explodecollections=True,
+            # gridsize=gridsize,
+            nb_parallel=nb_parallel,
+            batchsize=batchsize,
+            operation_prefix=f"{operation_name}-",
+        )
+
+        # Negative buffer to get back to the borders of the input geometries
+        # Use a larger mitre limit, otherwise there are a lot of small triangles that
+        # don't dissappear again.
+        step += 1
+        logger.info(f"Step {step} of {nb_steps}")
+        buff_diss_bufm_path = tempdir / "130_diss_bufp_diss_bufm.gpkg"
+        _geoops_gpd.buffer(
+            input_path=buff_diss_path,
+            output_path=buff_diss_bufm_path,
+            distance=-(distance / 2),
+            endcap_style=BufferEndCapStyle.SQUARE,
+            join_style=BufferJoinStyle.MITRE,
+            mitre_limit=2,
+            gridsize=gridsize,
+            nb_parallel=nb_parallel,
+            batchsize=batchsize,
+            operation_prefix=f"{operation_name}-",
+        )
+
+        # Determine which parts that were added were actually gaps within 'distance' in
+        # the original polygons, so they can be removed again.
+
+        # Determine all pieces added to the input in the process above.
+        step += 1
+        logger.info(f"Step {step} of {nb_steps}")
+        added_pieces_path = tempdir / "200_addedpieces.gpkg"
+        _geoops_sql.erase(
+            input_path=buff_diss_bufm_path,
+            erase_path=diss_path,
+            output_path=added_pieces_path,
+            explodecollections=True,
+            gridsize=gridsize,
+            nb_parallel=nb_parallel,
+            batchsize=batchsize,
+            operation_prefix=f"{operation_name}-",
+        )
+
+        # Build a filter to select the pieces that we want to erase again from the
+        # result because they were incorrectly added.
+        # The filter will depend on input parameters.
+        if close_internal_gaps:
+            # True, so also gaps in the original input boundaries should be closed.
+            # This means that in theory all pieces can be retained, but in practice
+            # there are some cases where the above algorithm adds unwanted area, so that
+            # needs to be erased again.
+            #
+            # Parameters that indicate that added pieces won't need to be erased:
+            #   - large areas (>= distance²) seem OK.
+            #   - if > 1 neighbour, seems OK.
+            #
+            # For all pieces that don't comply to the above, the following parameters
+            # indicate that they need to be selected to erase them:
+            #   - pieces can be very narrow slivers. E.g. alongside a long boundary with
+            #     a small bend, probably due to rounding side effects in the +/- buffer.
+            #   - pieces can be spikes. E.g. when a "road" of ~ 'distance' width is not
+            #     filled up between two input geometries has a bend. Depending on the
+            #     angle, the mitre of the negative buffer can leave a spike in place.
+            pieces_to_erase_filter = f"""
+                neighbours_count_distinct <= 1
+                AND geom_area < {distance} * {distance}
+                AND neighbours_perimeter/2 + neighbours_length <= 0.8 * geom_perimeter
+            """
+        else:
+            # False, so we want to keep all pieces that intersect with only 1 neighbour
+            # in the input, so they can be remove again from the result.
+            pieces_to_erase_filter = "neighbours_count_distinct <= 1"
+
+        # Notes:
+        # - The conversion to json followed by extraction from json allows to use a
+        #   correlated subquery to return multiple columns. Joining the subquery gives
+        #   very bad performance.
+        # - Every level of nesting of SQL queries is needed to get good performance, in
+        #   combination with "LIMIT -1 OFFSET 0" to avoid the subquery flattening.
+        #   Flattening e.g. "geom IS NOT NULL" leads to geom operation to be calculated
+        #   twice!
+        input1_layer_rtree = "rtree_{input1_layer}_{input1_geometrycolumn}"
+        input2_layer_rtree = "rtree_{input2_layer}_{input2_geometrycolumn}"
+        sql_stmt = f"""
+            WITH neighbours AS (
+                SELECT layer1_sub.rowid AS layer1_rowid
+                      ,layer2_sub.rowid AS layer2_rowid
+                      ,ST_Intersection(
+                          layer1_sub.{{input1_geometrycolumn}},
+                          layer2_sub.{{input2_geometrycolumn}}
+                       ) AS intersect_geom
+                  FROM {{input1_databasename}}."{{input1_layer}}" layer1_sub
+                  JOIN {{input1_databasename}}."{input1_layer_rtree}" layer1tree
+                    ON layer1_sub.rowid = layer1tree.id
+                  JOIN {{input2_databasename}}."{{input2_layer}}" layer2_sub
+                  JOIN {{input2_databasename}}."{input2_layer_rtree}" layer2tree
+                    ON layer2_sub.rowid = layer2tree.id
+                 WHERE 1=1
+                   AND layer1tree.minx <= layer2tree.maxx
+                   AND layer1tree.maxx >= layer2tree.minx
+                   AND layer1tree.miny <= layer2tree.maxy
+                   AND layer1tree.maxy >= layer2tree.miny
+                   AND ST_Intersects(
+                          layer1_sub.{{input1_geometrycolumn}},
+                          layer2_sub.{{input2_geometrycolumn}}) = 1
+              )
+            SELECT * FROM (
+              SELECT geom
+                    ,ST_Perimeter(geom) AS geom_perimeter
+                    ,ST_Area(geom) AS geom_area
+                    ,neighbours_json ->> '$.nb_distinct' AS neighbours_count_distinct
+                    ,neighbours_json ->> '$.length' AS neighbours_length
+                    ,neighbours_json ->> '$.perimeter' AS neighbours_perimeter
+                FROM (
+                  SELECT layer1.{{input1_geometrycolumn}} AS geom
+                        ,( SELECT json_object(
+                                    'nb_distinct', COUNT(DISTINCT layer2_rowid),
+                                    'length', SUM(ST_Length(intersect_geom)),
+                                    'perimeter', SUM(ST_Perimeter(intersect_geom))
+                                  )
+                             FROM neighbours
+                            WHERE neighbours.layer1_rowid = layer1.rowid
+                              AND neighbours.intersect_geom IS NOT NULL
+                            GROUP BY neighbours.layer1_rowid
+                            LIMIT -1 OFFSET 0
+                         ) AS neighbours_json
+                    FROM {{input1_databasename}}."{{input1_layer}}" layer1
+                   WHERE 1=1
+                     {{batch_filter}}
+                   LIMIT -1 OFFSET 0
+                  )
+                  LIMIT -1 OFFSET 0
+               )
+             WHERE geom IS NOT NULL
+               AND ({pieces_to_erase_filter})
+        """
+
+        step += 1
+        logger.info(f"Step {step} of {nb_steps}")
+        added_pieces_to_be_erased_input = tempdir / "210_addedpieces_to_be_erased.gpkg"
+        _geoops_sql.select_two_layers(
+            input1_path=added_pieces_path,
+            input2_path=input_path,
+            output_path=added_pieces_to_be_erased_input,
+            sql_stmt=sql_stmt,
+            input2_layer=input_layer,
+            # gridsize=gridsize,
+            nb_parallel=nb_parallel,
+            batchsize=batchsize,
+            operation_prefix=f"{operation_name}-",
+        )
+
+        step += 1
+        logger.info(f"Step {step} of {nb_steps}")
+        _geoops_sql.erase(
+            input_path=buff_diss_bufm_path,
+            erase_path=added_pieces_to_be_erased_input,
+            output_path=output_path,
+            output_layer=output_layer,
+            explodecollections=True,
+            gridsize=gridsize,
+            nb_parallel=nb_parallel,
+            batchsize=batchsize,
+            force=force,
+            operation_prefix=f"{operation_name}-",
+        )
+
+    finally:
+        shutil.rmtree(tempdir, ignore_errors=True)
+
+    logger.info(f"Ready, took {datetime.now()-start_time}")
 
 
 def apply(
@@ -175,8 +483,8 @@ def buffer(
               * MITRE: corners in the result are sharp
               * BEVEL: are flattened
         mitre_limit (float, optional): in case of join_style MITRE, if the
-            spiky result for a sharp angle becomes longer than this limit, it
-            is "beveled" at this distance. Defaults to 5.0.
+            spiky result for a sharp angle becomes longer than this ratio limit, it
+            is "beveled" using this maximum ratio. Defaults to 5.0.
         single_sided (bool, optional): only one side of the line is buffered,
             if distance is negative, the left side, if distance is positive,
             the right hand side. Only relevant for line geometries.