+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   (#262, #398)
 - Add `where_post` parameter to many two layer spatial operations (#312)
 - Add `where` parameter to `copy_layer` and `append_to` (#311)
+- Add `dissolve_within_distance` operation (#409)
 - Add support for lang+ algorithm in `simplify` (#334)
 - Add support to use `select` and `select_two_layers` on attribute tables (= tables
   without geometry column) and/or have an attribute table as result (#322, #379)